requirements.txt
alpaca-py
alpaca-trade-api>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0
streamlit>=1.28.0
plotly>=5.15.0
python-dotenv>=0.19.0
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
flask-compress>=1.13
pyjwt>=2.8.0
eventlet>=0.33.0
pandas-ta>=0.3.14b

# Performance optimization dependencies
aiohttp>=3.8.0
aiofiles>=23.0.0
websockets>=11.0.0
psutil>=5.9.0
redis>=4.5.0
aioredis>=2.0.0
uvloop>=0.17.0  # For better async performance on Unix
hiredis>=2.2.0  # For faster Redis operations


# Database and storage
# sqlite3 is part of Python standard library

# Enhanced logging and monitoring
colorlog>=6.7.0
python-json-logger>=2.0.0
prometheus-client>=0.16.0  # For metrics collection
statsd>=4.0.0  # For performance metrics

# Data validation and type checking
pydantic>=2.0.0
typing-extensions>=4.5.0

# Configuration management
pyyaml>=6.0
toml>=0.10.2

# Utility libraries
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0

# Advanced analytics and risk management
scipy>=1.10.0
scikit-learn>=1.3.0

# Machine Learning and AI
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
joblib>=1.3.0

# Testing and development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Educational and documentation
jupyter>=1.0.0
jupyterlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
ipywidgets>=8.0.0

# Performance profiling and monitoring
py-spy>=0.3.14  # For performance profiling
memory-profiler>=0.60.0  # For memory usage analysis
line-profiler>=4.0.0  # For line-by-line profiling
snakeviz>=2.1.1  # For cProfile visualization

# Additional performance libraries
cython>=3.0.0  # For potential Cython optimizations
numba>=0.57.0  # For JIT compilation of numerical code
fastapi>=0.68.0  # Alternative high-performance web framework (optional)
uvicorn>=0.18.0  # High-performance ASGI server
gunicorn>=20.1.0  # Production WSGI server with performance tuning
gevent>=22.10.0  # Alternative async library for better concurrency